Quanta emission change to 2 Quanta per minute
emission is currently close to 6 Quanta per minute and will be exponentially reduced to 0 in aproximately 200 years. Reducing the emission to 2 Quanta per minute will reduce the availability of Quanta.

lower emission rate will most likely increase demand of Quanta, the price, and people who are willing to point their mining rigs to qrl which will enlarge the network.